requirements_worker.txt
python-dotenv==1.0.0
pyrogram==2.0.106
requests==2.31.0
# اختر أحدهما حسب تفضيلك
pymongo==4.6.3  # MongoDB
redis==5.0.3    # Redis
# PDF Processing
pypdf>=4.0.0
python-docx>=0.8.11
reportlab>=4.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
pymupdf>=1.27.1
